feat(api): 新增刪除用戶 Prompt 歷史記錄的 API 端點
